refactor(ir): extract cycle-break algorithm into shared helper (Phase 1)
Pure relocation: the Tarjan SCC + synthetic-reg insertion algorithm
moves from `compiler/ir/trace_cycles.ts` into a standalone module
`compiler/ir/lowering/cycle_break.ts`. `trace_cycles.ts` becomes a
~20-line back-compat shim that calls the helper and returns just the
lowered program (matching the legacy signature consumed by
`strataPipeline`). `acyclic.ts` (added in Phase 2) now reuses the
helper's `findInstanceCycles` rather than carrying its own duplicate
implementation.

The new module exposes:
  - `findInstanceCycles(prog): InstanceDecl[][]` — pure detection.
    Reused by `acyclic.ts`'s `assertAcyclic` and by future
    realizations (iterative-fixed-point, WDF) that want the default
    Tarjan-based cycle detection.
  - `breakInstanceCycles(prog): { lowered, syntheticRegs, cycles }` —
    full cycle-break with provenance metadata. The metadata includes
    each broken SCC, its break-target instance, and the output ports
    that were promoted to synthetic regs. Future error formatters
    (Phase 4b's strict-policy diagnostics) consume this provenance
    to localize cycle violations.

Phase 3 will move the `breakInstanceCycles` call out of
`strataPipeline` and into the elaborator / session materializer;
Phase 5 will retire the `trace_cycles.ts` shim once nothing imports
it. This commit is a pure code move — diff is large but every change
is "move this hunk from A to B" with no behavior change.

Equivalence: 945 unit tests pass; 57 cross-backend tests pass.
